requirements.txt
psutil==5.7.2
PyQt5==5.15.0
pyserial==3.4
